requirements.txt
asgiref==3.11.0
charset-normalizer==3.4.4
Django==5.2.9
djangorestframework==3.16.1
mysqlclient==2.2.7
pillow==12.0.0
reportlab==4.4.7
sqlparse==0.5.5
tzdata==2025.3
